requirements.txt
python-telegram-bot==20.8
schedule==1.2.1
cryptography==42.0.5
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
apscheduler==3.10.4
requests==2.31.0
pytz==2024.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
